* commit 'ac8b3f3450f2d6c96061314f7580ce0353e81097':
  Only use double-precision floating point

+++ src/AlignSettings.h
@@ -20,9 +20,9 @@ struct AlignSettings : MM2Settings
     const std::vector<std::string> InputFiles;
 
     // alignment output filters
-    float MinPercConcordance;
-    float MinPercIdentity;
-    float MinPercIdentityGapComp;
+    double MinPercConcordance;
+    double MinPercIdentity;
+    double MinPercIdentityGapComp;
     int32_t MinAlignmentLength;
 
     const std::string SampleName;
